pyproject.toml
[project]
name = "obsidian-mcp-server"
version = "0.1.0"
description = "MCP server for accessing Obsidian notes via LiveSync CouchDB"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Obsidian MCP Server" }]
keywords = ["mcp", "obsidian", "notes", "couchdb", "livesync"]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "anyio>=4.5",
    "click>=8.1.0",
    "httpx>=0.27",
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-frontmatter>=1.0.0",
    "uvicorn>=0.24.0",
    "starlette>=0.27.0",
]

[project.scripts]
obsidian-mcp-server = "obsidian_mcp_server.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["obsidian_mcp_server"]

[tool.pyright]
include = ["obsidian_mcp_server"]
venvPath = "."
venv = ".venv"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.uv]
dev-dependencies = ["pyright>=1.1.378", "pytest>=8.3.3", "ruff>=0.6.9"] 